refactor(MemberService): separate MemberService interface
